backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
email-validator
